# services/qr-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Factorización QR de matrices rectangulares. Tras factorizar, envía Q y R a la Stats API (Node.js) y devuelve ambos resultados.\nSe expone a través de Kong en /api/qr y requiere un JWT (obtenlo con POST /api/auth/login).",
        "title": "QR API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/qr": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Recibe una matriz rectangular A (m×n), calcula A = Q·R (Q ortogonal m×m, R triangular superior m×n) y envía Q y R a la Stats API. Devuelve Q, R y las estadísticas.\nSi la misma matriz ya se calculó, responde desde el caché de Redis (cabecera X-Cache: HIT). Cada cálculo se guarda en el historial del usuario.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qr"
                ],
                "summary": "Factorización QR de una matriz",
                "parameters": [
                    {
                        "description": "Matriz a factorizar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpapi.FactorizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpapi.FactorizeResponse"
                        },
                        "headers": {
                            "X-Cache": {
                                "type": "string",
                                "description": "HIT si la respuesta salió del caché, MISS si se calculó"
                            }
                        }
                    },
                    "400": {
                        "description": "JSON o matriz inválidos",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token ausente, inválido o expirado (responde Kong)",
                        "schema": {
                            "$ref": "#/definitions/httpapi.GatewayError"
                        }
                    },
                    "415": {
                        "description": "El cuerpo no es application/json",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Límite de peticiones superado (responde Kong)",
                        "schema": {
                            "$ref": "#/definitions/httpapi.GatewayError"
                        }
                    },
                    "502": {
                        "description": "La Stats API no respondió correctamente",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/qr/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve las últimas factorizaciones del usuario del token, de la más reciente a la más antigua.\nCon scope=all (solo rol admin) devuelve las de todos los usuarios e incluye el autor de cada una.\nCon user=\u003cnombre\u003e (solo rol admin) devuelve las de ese usuario.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qr"
                ],
                "summary": "Historial de factorizaciones",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Cantidad máxima de resultados (1-50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "mine",
                            "all"
                        ],
                        "type": "string",
                        "description": "mine (por defecto) o all (solo admin)",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Historial de otro usuario (solo admin); no se combina con scope=all",
                        "name": "user",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpapi.HistoryResponse"
                        }
                    },
                    "400": {
                        "description": "scope inválido o combinado con user",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token ausente, inválido o expirado (responde Kong)",
                        "schema": {
                            "$ref": "#/definitions/httpapi.GatewayError"
                        }
                    },
                    "403": {
                        "description": "scope=all o el historial de otro usuario requieren el rol admin",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "El historial no está disponible",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/qr/usage": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cantidad de factorizaciones por usuario, cuántas salieron del caché de Redis y la fecha de la última. Solo para el rol admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qr"
                ],
                "summary": "Uso del servicio por usuario",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/history.Usage"
                        }
                    },
                    "401": {
                        "description": "Token ausente, inválido o expirado (responde Kong)",
                        "schema": {
                            "$ref": "#/definitions/httpapi.GatewayError"
                        }
                    },
                    "403": {
                        "description": "Requiere el rol admin",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "El historial no está disponible",
                        "schema": {
                            "$ref": "#/definitions/httpapi.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "history.Entry": {
            "type": "object",
            "properties": {
                "Q": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                },
                "R": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                },
                "cached": {
                    "description": "Cached indica si el resultado se obtuvo del caché de Redis.",
                    "type": "boolean",
                    "example": false
                },
                "columns": {
                    "type": "integer",
                    "example": 3
                },
                "createdAt": {
                    "type": "string",
                    "example": "2026-09-28T18:30:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 42
                },
                "matrix": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                },
                "rows": {
                    "type": "integer",
                    "example": 2
                },
                "statistics": {
                    "$ref": "#/definitions/stats.Statistics"
                },
                "username": {
                    "description": "Username es el autor del cálculo; solo se incluye en el historial de todos los usuarios.",
                    "type": "string",
                    "example": "analyst"
                }
            }
        },
        "history.Usage": {
            "type": "object",
            "properties": {
                "cacheHits": {
                    "type": "integer",
                    "example": 9
                },
                "total": {
                    "type": "integer",
                    "example": 30
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/history.UserUsage"
                    }
                }
            }
        },
        "history.UserUsage": {
            "type": "object",
            "properties": {
                "cacheHits": {
                    "type": "integer",
                    "example": 4
                },
                "count": {
                    "type": "integer",
                    "example": 12
                },
                "lastAt": {
                    "type": "string",
                    "example": "2026-09-28T18:30:00Z"
                },
                "username": {
                    "type": "string",
                    "example": "analyst"
                }
            }
        },
        "httpapi.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code es un identificador estable del tipo de error.",
                    "type": "string",
                    "example": "INVALID_MATRIX"
                },
                "message": {
                    "description": "Message explica el error en lenguaje natural.",
                    "type": "string",
                    "example": "la matriz debe ser rectangular: la fila 2 tiene 1 columnas y la fila 1 tiene 3"
                },
                "requestId": {
                    "description": "RequestID permite encontrar la petición en los logs.",
                    "type": "string",
                    "example": "3f1b6c2e-7a55-4a8e-9b1d-0c5e2f7a9d10"
                }
            }
        },
        "httpapi.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/httpapi.ErrorBody"
                }
            }
        },
        "httpapi.FactorizeRequest": {
            "type": "object",
            "properties": {
                "matrix": {
                    "description": "Matrix es la matriz rectangular A (m×n) a factorizar.",
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                }
            }
        },
        "httpapi.FactorizeResponse": {
            "type": "object",
            "properties": {
                "Q": {
                    "description": "Q es la matriz ortogonal (m×m).",
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                },
                "R": {
                    "description": "R es la matriz triangular superior (m×n), con diagonal no negativa.",
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                },
                "statistics": {
                    "description": "Statistics son las estadísticas calculadas por la Stats API sobre Q y R.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stats.Statistics"
                        }
                    ]
                }
            }
        },
        "httpapi.GatewayError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Unauthorized"
                }
            }
        },
        "httpapi.HistoryResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/history.Entry"
                    }
                }
            }
        },
        "stats.MatrixSummary": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "integer",
                    "example": 3
                },
                "isDiagonal": {
                    "type": "boolean",
                    "example": false
                },
                "rows": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "stats.Statistics": {
            "type": "object",
            "properties": {
                "average": {
                    "description": "Average es el promedio de todos los elementos de Q y R.",
                    "type": "number",
                    "example": 2.013046
                },
                "count": {
                    "description": "Count es la cantidad de elementos analizados.",
                    "type": "integer",
                    "example": 10
                },
                "isAnyDiagonal": {
                    "description": "IsAnyDiagonal indica si al menos una de las matrices es diagonal.",
                    "type": "boolean",
                    "example": false
                },
                "matrices": {
                    "description": "Matrices resume cada matriz por nombre (\"Q\" y \"R\").",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/stats.MatrixSummary"
                    }
                },
                "max": {
                    "description": "Max es el valor máximo entre todos los elementos de Q y R.",
                    "type": "number",
                    "example": 6.548462
                },
                "min": {
                    "description": "Min es el valor mínimo entre todos los elementos de Q y R.",
                    "type": "number",
                    "example": -0.242536
                },
                "sum": {
                    "description": "Sum es la suma de todos los elementos de Q y R.",
                    "type": "number",
                    "example": 20.130457
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Escribe \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}